# ENH, SIMD: Extend universal intrinsics to support IBMZ