--- a/numpy/core/src/common/simd/avx2/avx2.h
+++ b/numpy/core/src/common/simd/avx2/avx2.h
@@ -3,12 +3,14 @@
 #endif
 #define NPY_SIMD 256
 #define NPY_SIMD_WIDTH 32
+#define NPY_SIMD_F32 1
 #define NPY_SIMD_F64 1
 #ifdef NPY_HAVE_FMA3
     #define NPY_SIMD_FMA3 1 // native support
 #else
     #define NPY_SIMD_FMA3 0 // fast emulated
 #endif
+#define NPY_SIMD_BIGENDIAN 0
 // Enough limit to allow us to use _mm256_i32gather_*
 #define NPY_SIMD_MAXLOAD_STRIDE32 (0x7fffffff / 8)