--- a/numpy/core/src/common/simd/avx512/avx512.h
+++ b/numpy/core/src/common/simd/avx512/avx512.h
@@ -3,8 +3,10 @@
 #endif
 #define NPY_SIMD 512
 #define NPY_SIMD_WIDTH 64
+#define NPY_SIMD_F32 1
 #define NPY_SIMD_F64 1
 #define NPY_SIMD_FMA3 1 // native support
+#define NPY_SIMD_BIGENDIAN 0
 // Enough limit to allow us to use _mm512_i32gather_* and _mm512_i32scatter_*
 #define NPY_SIMD_MAXLOAD_STRIDE32  (0x7fffffff / 16)
 #define NPY_SIMD_MAXSTORE_STRIDE32 (0x7fffffff / 16)