--- a/numpy/core/src/common/simd/neon/neon.h
+++ b/numpy/core/src/common/simd/neon/neon.h
@@ -4,7 +4,7 @@
 
 #define NPY_SIMD 128
 #define NPY_SIMD_WIDTH 16
-
+#define NPY_SIMD_F32 1
 #ifdef __aarch64__
     #define NPY_SIMD_F64 1
 #else
@@ -15,6 +15,7 @@
 #else
     #define NPY_SIMD_FMA3 0  // HW emulated
 #endif
+#define NPY_SIMD_BIGENDIAN 0
 
 typedef uint8x16_t  npyv_u8;
 typedef int8x16_t   npyv_s8;