… doc/release/upcoming_changes/20913.improvement.rst
… numpy/core/src/common/simd/emulate_maskop.h
… numpy/core/src/common/simd/vec/conversion.h
… numpy/core/src/common/simd/vec/math.h
… numpy/core/src/common/simd/vec/memory.h
… numpy/core/src/common/simd/vec/misc.h
… numpy/core/src/common/simd/vec/operators.h
… numpy/core/src/common/simd/vec/reorder.h
… numpy/core/src/common/simd/vec/utils.h
… numpy/core/src/common/simd/vec/vec.h
… numpy/core/src/multiarray/argfunc.dispatch.c.src
… numpy/core/src/umath/loops_arithm_fp.dispatch.c.src
… numpy/core/src/umath/loops_comparison.dispatch.c.src
… numpy/core/src/umath/loops_hyperbolic.dispatch.c.src
… numpy/core/src/umath/loops_minmax.dispatch.c.src
… numpy/core/src/umath/loops_trigonometric.dispatch.c.src
… numpy/core/src/umath/loops_unary_fp.dispatch.c.src
… numpy/core/tests/test_simd.py
… tools/travis-test.sh