#### Extend universal intrinsics to support IBMZ

It covers SIMD operations for all datatypes starting
from z/Arch11 a.k.a IBM Z13, except for single-precision
which requires minimum z/Arch12 a.k.a IBMZ 14 to be dispatched.

This patch rename the branch /simd/vsx to /simd/vec, the new
the path holds the definitions of universal intrinsics for
both Power and Z architectures.

This patch also adds new preprocessor identifiers:

  - NPY_SIMD_BIGENDIAN: 1 if the enabled SIMD extension
  is running on big-endian mode otherwise 0.

  - NPY_SIMD_F32: 1 if the enabled SIMD extension
  supports single-precision otherwise 0.


TODO:
- [x] release note
- [x] benchmark 

### Benchmark
The following benchmark is inferential and does not accurately reflect the true change, since it was accomplished using an unstable VM. 


<details>
<summary>CPU</summary>

```Bash
Architecture:                    s390x
CPU op-mode(s):                  32-bit, 64-bit
Byte Order:                      Big Endian
CPU(s):                          2
On-line CPU(s) list:             0,1
Thread(s) per core:              1
Core(s) per socket:              1
Socket(s) per book:              1
Book(s) per drawer:              1
Drawer(s):                       2
NUMA node(s):                    1
Vendor ID:                       IBM/S390
Machine type:                    8561
CPU dynamic MHz:                 5200
CPU static MHz:                  5200
BogoMIPS:                        3241.00
Hypervisor:                      z/VM 7.1.0
Hypervisor vendor:               IBM
Virtualization type:             full
Dispatching mode:                horizontal
L1d cache:                       256 KiB
L1i cache:                       256 KiB
L2d cache:                       8 MiB
L2i cache:                       8 MiB
L3 cache:                        256 MiB
L4 cache:                        960 MiB
NUMA node0 CPU(s):               0,1
Vulnerability Itlb multihit:     Not affected
Vulnerability L1tf:              Not affected
Vulnerability Mds:               Not affected
Vulnerability Meltdown:          Not affected
Vulnerability Spec store bypass: Not affected
Vulnerability Spectre v1:        Mitigation; __user pointer sanitization
Vulnerability Spectre v2:        Mitigation; etokens
Vulnerability Srbds:             Not affected
Vulnerability Tsx async abort:   Not affected
Flags:                           esan3 zarch stfle msa ldisp eimm dfp edat etf3eh highgprs te
                                 vx vxd vxe gs vxe2 vxp sort dflt sie
```
</details>

<details>
<summary>OS</summary>

```Bash
Linux numpy 5.4.0-104-generic #118-Ubuntu SMP Wed Mar 2 19:02:13 UTC 2022 s390x s390x s390x GNU/Linux
Python 3.8.10
gcc (Ubuntu 11.1.0-1ubuntu1~20.04) 11.1.0
```
</details>

#### Benchmark
<details>
 <summary>VXE2</summary>

```Bash
unset NPY_DISABLE_CPU_FEATURES
python runtests.py --bench-compare parent/main
```
```Bash
    before           after         ratio
     [982fcd38]       [47d54c6d]
     <zsystem_sup~5>       <zsystem_sup>
+     8.41±0.03μs       10.2±0.3μs     1.21  bench_function_base.Sort.time_sort('merge', 'int64', ('uniform',))
+      82.8±0.4μs       99.0±0.3μs     1.19  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 1, 1, 2, 'h')
+      83.0±0.3μs       99.2±0.3μs     1.19  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 2, 1, 2, 'h')
+      83.5±0.7μs       99.7±0.7μs     1.19  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 1, 1, 4, 'h')
+      84.3±0.8μs        101±0.5μs     1.19  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 1, 1, 'h')
+      83.1±0.2μs       99.2±0.2μs     1.19  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 2, 1, 2, 'h')
+      85.0±0.7μs        101±0.4μs     1.19  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 1, 2, 'h')
+      83.1±0.4μs       99.0±0.3μs     1.19  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 2, 2, 1, 'h')
+     10.3±0.03μs      12.2±0.06μs     1.19  bench_function_base.Sort.time_sort('merge', 'float32', ('ordered',))
+      83.7±0.5μs       9.2±0.4μs     1.19  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 1, 1, 2, 'h')
+       127±0.4μs          149±3μs     1.17  bench_function_base.Sort.time_argsort('quick', 'int32', ('reversed',))
+         322±1μs          364±2μs     1.13  bench_function_base.Sort.time_argsort('quick', 'int64', ('sorted_block', 1000))
+        87.6±3μs       99.0±0.8μs     1.13  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 1, 4, 4, 'b')
+         249±1μs          281±2μs     1.13  bench_core.CountNonzero.time_count_nonzero(3, 10000, <class 'str'>)
+      87.4±0.8μs       98.6±0.2μs     1.13  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 4, 2, 'b')
+         131±5μs          148±2μs     1.13  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 1, 2, 'q')
+     11.9±0.08μs       13.4±0.3μs     1.13  bench_ufunc_strides.AVX_cmplx_arithmetic.time_ufunc('multiply', 1, 'D')
+      87.6±0.5μs       98.5±0.2μs     1.12  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 2, 2, 4, 'b')
+      13.3±0.2μs       14.9±0.2μs     1.12  bench_ufunc_strides.AVX_cmplx_arithmetic.time_ufunc('multiply', 2, 'D')
+     1.47±0.02ms      1.64±0.05ms     1.12  bench_lib.Pad.time_pad((256, 128, 1), 8, 'edge')
+       167±0.9μs          186±4μs     1.11  bench_core.CountNonzero.time_count_nonzero(2, 10000, <class 'str'>)
+         113±3μs          126±2μs     1.11  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 1, 1, 'Q')
+      5.71±0.2ms       6.32±0.2ms     1.11  bench_lib.Pad.time_pad((4194304,), 1, 'mean')
+         283±7μs         312±10μs     1.10  bench_core.PackBits.time_packbits_axis0(<class 'bool'>)
+         329±3μs          363±2μs     1.10  bench_function_base.Sort.time_argsort('quick', 'int32', ('sorted_block', 1000))
+      70.6±0.5μs       77.9±0.7μs     1.10  bench_ufunc.CustomArrayFloorDivideInt.time_floor_divide_int(<class 'numpy.int64'>, 10000)
+         114±1μs          126±3μs     1.10  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 1, 1, 'L')
+      87.7±0.9μs       96.6±0.2μs     1.10  bench_function_base.Sort.time_sort('merge', 'float64', ('sorted_block', 100))
+       299±0.8μs          329±1μs     1.10  bench_function_base.Sort.time_sort('quick', 'int16', ('sorted_block', 1000))
+      79.2±0.2μs       87.1±0.1μs     1.10  bench_function_base.Where.time_interleaved_zeros_x8
+      61.0±0.5μs       67.0±0.5μs     1.10  bench_function_base.Select.time_select
+         117±5μs          129±3μs     1.10  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 1, 1, 'q')
+      82.3±0.4μs         89.9±2μs     1.09  bench_core.CountNonzero.time_count_nonzero(1, 10000, <class 'str'>)
+      52.9±0.6μs       57.5±0.3μs     1.09  bench_function_base.Sort.time_argsort('merge', 'int16', ('sorted_block', 10))
+         195±1μs        212±0.6μs     1.09  bench_function_base.Sort.time_sort('merge', 'int32', ('sorted_block', 10))
+     3.33±0.04ms       3.60±0.3ms     1.08  bench_lib.Pad.time_pad((1, 1, 1, 1, 1), 8, 'edge')
+      7.01±0.1ms       7.57±0.2ms     1.08  bench_ufunc.CustomArrayFloorDivideInt.time_floor_divide_int(<class 'numpy.int64'>, 1000000)
+      5.26±0.1ms       5.68±0.2ms     1.08  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 10000)
+         138±4μs          148±3μs     1.08  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 1, 2, 'L')
+      20.5±0.4ms       22.0±0.3ms     1.07  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 100000)
+         408±2μs        438±0.8μs     1.07  bench_function_base.Sort.time_argsort('quick', 'int64', ('sorted_block', 100))
+         184±2μs          197±8μs     1.07  bench_lib.Pad.time_pad((4, 4, 4, 4), 8, 'constant')
+        52.7±2ms         56.4±2ms     1.07  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 100000)
+         121±2μs          129±3μs     1.07  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 1, 1, 'L')
+       104±0.7μs        111±0.3μs     1.07  bench_function_base.Sort.time_sort('quick', 'float32', ('uniform',))
+       368±0.7μs          393±1μs     1.07  bench_function_base.Sort.time_sort('quick', 'int16', ('sorted_block', 100))
+         510±5μs          544±3μs     1.06  bench_function_base.Sort.time_argsort('heap', 'float64', ('ordered',))
+     1.80±0.02ms      1.91±0.03ms     1.06  bench_core.CountNonzero.time_count_nonzero(3, 1000000, <class 'numpy.int64'>)
+     3.44±0.03ms      3.65±0.07ms     1.06  bench_core.CountNonzero.time_count_nonzero_axis(2, 1000000, <class 'numpy.int64'>)
+         534±2μs         566±10μs     1.06  bench_ufunc_strides.Unary.time_ufunc(<ufunc 'exp2'>, 4, 4, 'd')
+      49.3±0.4μs       52.3±0.4μs     1.06  bench_function_base.Sort.time_sort('merge', 'float32', ('sorted_block', 1000))
+         141±4μs          149±3μs     1.06  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 2, 1, 'l')
+         121±3μs          129±4μs     1.06  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 1, 1, 'q')
+      93.3±0.7μs         98.8±1μs     1.06  bench_function_base.Sort.time_argsort('merge', 'int64', ('sorted_block', 100))
+        942±30μs         997±20μs     1.06  bench_reduce.AddReduceSeparate.time_reduce(1, 'int16')
+         410±6μs        434±0.9μs     1.06  bench_function_base.Sort.time_argsort('quick', 'int32', ('sorted_block', 100))
+         531±2μs          561±2μs     1.06  bench_function_base.Sort.time_argsort('heap', 'float32', ('ordered',))
+         526±1μs         556±10μs     1.06  bench_function_base.Sort.time_argsort('merge', 'int64', ('random',))
+        75.5±1μs         79.8±2μs     1.06  bench_function_base.Sort.time_argsort('quick', 'uint32', ('ordered',))
+      57.0±0.7μs       60.2±0.4μs     1.06  bench_function_base.Sort.time_argsort('merge', 'int16', ('sorted_block', 100))
+         534±7μs          562±4μs     1.05  bench_ufunc_strides.Unary.time_ufunc(<ufunc 'exp2'>, 4, 2, 'd')
+       727±0.9μs         765±30μs     1.05  bench_ufunc_strides.Unary.time_ufunc(<ufunc 'log'>, 2, 2, 'd')
+        683±10μs          719±4μs     1.05  bench_lib.Pad.time_pad((256, 128, 1), 8, 'constant')
+        46.0±2μs       48.4±0.2μs     1.05  bench_function_base.Sort.time_argsort('heap', 'int64', ('uniform',))
+         121±1μs          128±4μs     1.05  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 1, 1, 'Q')
+     2.10±0.02ms      2.21±0.03ms     1.05  bench_lib.Pad.time_pad((256, 128, 1), (0, 32), 'mean')
+         237±1ms          249±2ms     1.05  bench_function_base.Histogram2D.time_fine_binning
+     1.20±0.01ms      1.26±0.01ms     1.05  bench_core.CountNonzero.time_count_nonzero(2, 1000000, <class 'numpy.int64'>)
-      5.47±0.1μs      5.21±0.05μs     0.95  bench_ma.Indexing.time_1d(False, 1, 10)
-     1.50±0.02ms      1.43±0.01ms     0.95  bench_lib.Nan.time_nanargmin(200000, 50.0)
-      30.5±0.1μs       29.1±0.2μs     0.95  bench_function_base.Median.time_odd_inplace
-      88.8±0.7μs       84.6±0.4μs     0.95  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 1, 4, 1, 'H')
-      13.7±0.1μs      13.1±0.08μs     0.95  bench_ma.UFunc.time_2d(False, False, 10)
-         696±4μs        662±0.5μs     0.95  bench_function_base.Sort.time_sort('heap', 'float32', ('sorted_block', 10))
-      36.9±0.1μs       35.1±0.1μs     0.95  bench_core.CountNonzero.time_count_nonzero_axis(2, 10000, <class 'numpy.int8'>)
-      72.6±0.6μs         69.0±1μs     0.95  bench_ufunc.CustomArrayFloorDivideInt.time_floor_divide_int(<class 'numpy.int16'>, 10000)
-      90.9±0.6μs       86.3±0.8μs     0.95  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 1, 4, 4, 'H')
-      88.3±0.3μs       83.9±0.4μs     0.95  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 2, 1, 2, 'H')
-       473±0.7μs          449±1μs     0.95  bench_function_base.Sort.time_argsort('quick', 'int16', ('sorted_block', 10))
-      88.3±0.3μs       83.8±0.5μs     0.95  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 1, 2, 2, 'H')
-      38.8±0.9μs      36.8±0.07μs     0.95  bench_linalg.Linalg.time_op('norm', 'complex64')
-      88.9±0.7μs         84.2±1μs     0.95  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 1, 1, 4, 'H')
-      92.0±0.1μs         87.1±1μs     0.95  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 4, 2, 'H')
-      5.79±0.1μs      5.48±0.01μs     0.95  bench_ma.Indexing.time_1d(True, 1, 1000)
-      87.2±0.3μs       82.5±0.4μs     0.95  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 1, 2, 1, 'H')
-     7.23±0.05ms      6.85±0.03ms     0.95  bench_ufunc.CustomArrayFloorDivideInt.time_floor_divide_int(<class 'numpy.int16'>, 1000000)
-     2.34±0.02ms      2.21±0.02ms     0.95  bench_lib.Nan.time_nanstd(200000, 90.0)
-      5.46±0.1μs      5.17±0.05μs     0.95  bench_ma.Indexing.time_1d(False, 2, 10)
-      5.46±0.1μs       5.16±0.1μs     0.95  bench_ma.Indexing.time_1d(False, 2, 1000)
-     5.86±0.05μs      5.54±0.02μs     0.95  bench_indexing.ScalarIndexing.time_assign(0)
-      5.45±0.1μs       5.15±0.1μs     0.95  bench_ma.Indexing.time_0d(False, 2, 10)
-      5.80±0.1μs      5.48±0.04μs     0.94  bench_ma.Indexing.time_0d(True, 1, 10)
-      88.5±0.6μs       83.6±0.5μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 2, 1, 2, 'H')
-      87.7±0.4μs       82.8±0.5μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 1, 2, 1, 'H')
-      5.83±0.1μs      5.50±0.04μs     0.94  bench_ma.Indexing.time_0d(True, 1, 100)
-      90.6±0.3μs       85.5±0.8μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 2, 2, 'H')
-       151±0.8μs        143±0.5μs     0.94  bench_function_base.Sort.time_argsort('quick', 'int16', ('reversed',))
-      89.5±0.5μs       84.4±0.5μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 2, 4, 1, 'H')
-     2.33±0.02ms      2.20±0.01ms     0.94  bench_lib.Nan.time_nanvar(200000, 90.0)
-      87.8±0.3μs       82.8±0.4μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 1, 1, 2, 'H')
-      18.6±0.2μs       17.5±0.2μs     0.94  bench_ufunc_strides.AVX_cmplx_arithmetic.time_ufunc('multiply', 4, 'D')
-      92.8±0.9μs         87.4±1μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 4, 2, 'H')
-      89.9±0.5μs       84.7±0.6μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 2, 1, 'H')
-      90.1±0.4μs       84.8±0.5μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 1, 2, 'H')
-      89.9±0.4μs       84.6±0.5μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 2, 4, 1, 'H')
-      5.73±0.1μs      5.40±0.03μs     0.94  bench_ma.Indexing.time_1d(True, 1, 100)
-        89.3±1μs         83.9±1μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 1, 1, 4, 'H')
-      91.4±0.5μs       85.9±0.8μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 4, 1, 'H')
-      5.87±0.1μs      5.52±0.05μs     0.94  bench_ma.Indexing.time_0d(True, 2, 10)
-      78.0±0.8μs       73.2±0.2μs     0.94  bench_function_base.Sort.time_argsort('quick', 'int32', ('ordered',))
-        90.1±1μs       84.7±0.7μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 1, 4, 1, 'H')
-      88.1±0.4μs       82.7±0.2μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 1, 2, 2, 'H')
-         817±4μs          767±2μs     0.94  bench_ufunc.UFunc.time_ufunc_types('reciprocal')
-     1.42±0.01ms      1.33±0.01ms     0.94  bench_lib.Nan.time_nanvar(200000, 0.1)
-      90.2±0.3μs       84.6±0.5μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 1, 2, 'H')
-         491±3μs          461±2μs     0.94  bench_function_base.Sort.time_argsort('quick', 'int16', ('sorted_block', 100))
-        91.9±1μs         86.1±1μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 2, 2, 4, 'H')
-      5.89±0.2μs      5.52±0.06μs     0.94  bench_ma.Indexing.time_1d(True, 2, 100)
-      5.49±0.2μs      5.15±0.05μs     0.94  bench_ma.Indexing.time_1d(False, 1, 100)
-        91.4±1μs       85.6±0.2μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 1, 4, 4, 'H')
-         355±6μs          332±2μs     0.94  bench_ufunc.UFunc.time_ufunc_types('square')
-      87.8±0.5μs       82.2±0.2μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 2, 2, 1, 'H')
-      90.2±0.8μs       84.5±0.4μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 2, 1, 4, 'H')
-     5.46±0.05μs       5.11±0.1μs     0.94  bench_ma.Indexing.time_0d(False, 1, 100)
-     1.50±0.01ms      1.41±0.01ms     0.94  bench_lib.Nan.time_nanstd(200000, 2.0)
-     5.82±0.04μs      5.44±0.02μs     0.94  bench_ma.Indexing.time_0d(True, 1, 1000)
-      87.6±0.3μs       81.9±0.2μs     0.94  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 1, 1, 2, 'H')
-      89.7±0.2μs       83.8±0.3μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 2, 1, 'H')
-     87.6±0.09μs       81.9±0.1μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 2, 1, 1, 'H')
-      91.2±0.9μs       85.2±0.9μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 2, 4, 2, 'H')
-      90.5±0.9μs       84.5±0.8μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 2, 4, 2, 'H')
-      91.2±0.7μs         85.2±1μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 2, 2, 4, 'H')
-     5.46±0.07μs      5.10±0.09μs     0.93  bench_ma.Indexing.time_0d(False, 2, 1000)
-     1.42±0.01ms      1.33±0.02ms     0.93  bench_lib.Nan.time_nanvar(200000, 0)
-      92.3±0.5μs       86.1±0.7μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 2, 4, 'H')
-     5.50±0.07μs      5.13±0.08μs     0.93  bench_ma.Indexing.time_0d(False, 2, 100)
-      91.3±0.3μs       85.1±0.6μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 1, 4, 'H')
-        89.8±1μs       83.7±0.3μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 2, 2, 2, 'H')
-         466±4μs          434±1μs     0.93  bench_function_base.Sort.time_sort('merge', 'uint32', ('random',))
-         277±3μs          258±1μs     0.93  bench_ufunc_strides.Unary.time_ufunc(<ufunc 'reciprocal'>, 2, 1, 'd')
-      90.9±0.5μs       84.6±0.3μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 2, 2, 'H')
-      94.1±0.8μs       87.6±0.6μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 4, 4, 'H')
-      89.1±0.6μs       82.9±0.3μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 2, 2, 2, 'H')
-      17.8±0.2μs      16.5±0.09μs     0.93  bench_core.VarComplex.time_var(1000)
-         275±3μs          256±3μs     0.93  bench_ufunc_strides.Unary.time_ufunc(<ufunc 'reciprocal'>, 4, 1, 'd')
-       273±0.5μs          254±2μs     0.93  bench_ufunc_strides.Unary.time_ufunc(<ufunc 'reciprocal'>, 4, 2, 'd')
-      90.0±0.2μs       83.6±0.5μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 2, 1, 4, 'H')
-      89.4±0.6μs       83.1±0.5μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 1, 1, 'H')
-      93.0±0.7μs       86.4±0.6μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('maximum', 4, 2, 4, 'H')
-      89.9±0.4μs       83.4±0.5μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 1, 1, 'H')
-         272±2μs          253±5μs     0.93  bench_ufunc_strides.Unary.time_ufunc(<ufunc 'reciprocal'>, 2, 2, 'd')
-     1.42±0.01ms      1.32±0.01ms     0.93  bench_lib.Nan.time_nanstd(200000, 0.1)
-      88.4±0.4μs       82.0±0.3μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 2, 1, 1, 'H')
-      5.95±0.1μs      5.52±0.08μs     0.93  bench_ma.Indexing.time_1d(True, 2, 10)
-      91.4±0.5μs       84.8±0.3μs     0.93  bench_ufunc_strides.BinaryInt.time_ufunc('minimum', 4, 1, 4, 'H')
…